small refactor for get_logs()
